use uint32 only for storage

+++ contracts/modules/CollectionOffers/V1/CollectionOfferBookV1.sol
@@ -32,7 +32,7 @@ contract CollectionOfferBookV1 {
 
     /// @notice The floor offer ID for a given collection
     /// @dev ERC-721 token address => Floor offer ID
-    mapping(address => uint32) public floorOfferId;
+    mapping(address => uint256) public floorOfferId;
 
     /// @notice The floor offer amount for a given collection
     /// @dev ERC-721 token address => Floor offer amount
@@ -40,7 +40,7 @@ contract CollectionOfferBookV1 {
 
     /// @notice The ceiling offer ID for a given collection
     /// @dev ERC-721 token address => Ceiling offer ID
-    mapping(address => uint32) public ceilingOfferId;
+    mapping(address => uint256) public ceilingOfferId;
 
     /// @notice The ceiling offer amount for a given collection
     /// @dev ERC-721 token address => Ceiling offer amount
@@ -56,7 +56,7 @@ contract CollectionOfferBookV1 {
         address _collection,
         uint256 _amount,
         address _maker
-    ) internal returns (uint32) {
+    ) internal returns (uint256) {
         unchecked {
             ++offerCount;
         }
@@ -73,20 +73,20 @@ contract CollectionOfferBookV1 {
 
             // Else if offer is greater than current ceiling, mark as new ceiling
         } else if (_isNewCeiling(_collection, _amount)) {
-            uint32 prevCeilingId = ceilingOfferId[_collection];
+            uint256 prevCeilingId = ceilingOfferId[_collection];
 
-            offers[_collection][prevCeilingId].nextId = offerCount;
-            offers[_collection][offerCount] = Offer({maker: _maker, amount: _amount, id: offerCount, prevId: prevCeilingId, nextId: 0});
+            offers[_collection][prevCeilingId].nextId = uint32(offerCount);
+            offers[_collection][offerCount] = Offer({maker: _maker, amount: _amount, id: offerCount, prevId: uint32(prevCeilingId), nextId: 0});
 
             ceilingOfferId[_collection] = offerCount;
             ceilingOfferAmount[_collection] = _amount;
 
             // Else if offer is less than or equal to current floor, mark as new floor
         } else if (_isNewFloor(_collection, _amount)) {
-            uint32 prevFloorId = floorOfferId[_collection];
+            uint256 prevFloorId = floorOfferId[_collection];
 
-            offers[_collection][prevFloorId].prevId = offerCount;
-            offers[_collection][offerCount] = Offer({maker: _maker, amount: _amount, id: offerCount, prevId: 0, nextId: prevFloorId});
+            offers[_collection][prevFloorId].prevId = uint32(offerCount);
+            offers[_collection][offerCount] = Offer({maker: _maker, amount: _amount, id: offerCount, prevId: 0, nextId: uint32(prevFloorId)});
 
             floorOfferId[_collection] = offerCount;
             floorOfferAmount[_collection] = _amount;
@@ -105,8 +105,8 @@ contract CollectionOfferBookV1 {
             offers[_collection][offerCount] = Offer({maker: _maker, amount: _amount, id: offerCount, prevId: offer.prevId, nextId: offer.id});
 
             // Update neighboring pointers
-            offers[_collection][offer.id].prevId = offerCount;
-            offers[_collection][offer.prevId].nextId = offerCount;
+            offers[_collection][offer.id].prevId = uint32(offerCount);
+            offers[_collection][offer.prevId].nextId = uint32(offerCount);
         }
 
         return offerCount;
@@ -121,7 +121,7 @@ contract CollectionOfferBookV1 {
     function _updateOffer(
         Offer storage _offer,
         address _collection,
-        uint32 _offerId,
+        uint256 _offerId,
         uint256 _newAmount,
         bool _increase
     ) internal {
@@ -154,18 +154,18 @@ contract CollectionOfferBookV1 {
             // Else if offer is new ceiling --
         } else if (_isNewCeiling(_collection, _newAmount)) {
             // Get previous neighbors
-            uint32 prevId = _offer.prevId;
-            uint32 nextId = _offer.nextId;
+            uint256 prevId = _offer.prevId;
+            uint256 nextId = _offer.nextId;
 
             // Update previous neighbors
             _connectNeighbors(_collection, _offerId, prevId, nextId);
 
             // Update previous ceiling
-            uint32 prevCeilingId = ceilingOfferId[_collection];
-            offers[_collection][prevCeilingId].nextId = _offerId;
+            uint256 prevCeilingId = ceilingOfferId[_collection];
+            offers[_collection][prevCeilingId].nextId = uint32(_offerId);
 
             // Update offer to be new ceiling
-            _offer.prevId = prevCeilingId;
+            _offer.prevId = uint32(prevCeilingId);
             _offer.nextId = 0;
             _offer.amount = _newAmount;
 
@@ -176,18 +176,18 @@ contract CollectionOfferBookV1 {
             // Else if offer is new floor --
         } else if (_isNewFloor(_collection, _newAmount)) {
             // Get previous neighbors
-            uint32 prevId = _offer.prevId;
-            uint32 nextId = _offer.nextId;
+            uint256 prevId = _offer.prevId;
+            uint256 nextId = _offer.nextId;
 
             // Update previous neighbors
             _connectNeighbors(_collection, _offerId, prevId, nextId);
 
             // Update previous floor
-            uint32 prevFloorId = floorOfferId[_collection];
-            offers[_collection][prevFloorId].prevId = _offerId;
+            uint256 prevFloorId = floorOfferId[_collection];
+            offers[_collection][prevFloorId].prevId = uint32(_offerId);
 
             // Update offer to be new floor
-            _offer.nextId = prevFloorId;
+            _offer.nextId = uint32(prevFloorId);
             _offer.prevId = 0;
             _offer.amount = _newAmount;
 
@@ -215,7 +215,7 @@ contract CollectionOfferBookV1 {
     /// @notice Removes an offer from its collection's offer book
     /// @param _collection The address of the ERC-721 collection
     /// @param _offerId The ID of the offer
-    function _removeOffer(address _collection, uint32 _offerId) internal {
+    function _removeOffer(address _collection, uint256 _offerId) internal {
         // If offer is only one for collection, remove all associated data
         if (_isOnlyOffer(_collection, _offerId)) {
             delete floorOfferId[_collection];
@@ -226,7 +226,7 @@ contract CollectionOfferBookV1 {
 
             // Else if the offer is the current floor, update the collection's floor before removing
         } else if (_isFloorOffer(_collection, _offerId)) {
-            uint32 newFloorId = offers[_collection][_offerId].nextId;
+            uint256 newFloorId = offers[_collection][_offerId].nextId;
             uint256 newFloorAmount = offers[_collection][newFloorId].amount;
 
             offers[_collection][newFloorId].prevId = 0;
@@ -238,7 +238,7 @@ contract CollectionOfferBookV1 {
 
             // Else if the offer is the current ceiling, update the collection's ceiling before removing
         } else if (_isCeilingOffer(_collection, _offerId)) {
-            uint32 newCeilingId = offers[_collection][_offerId].prevId;
+            uint256 newCeilingId = offers[_collection][_offerId].prevId;
             uint256 newCeilingAmount = offers[_collection][newCeilingId].amount;
 
             offers[_collection][newCeilingId].nextId = 0;
@@ -252,8 +252,8 @@ contract CollectionOfferBookV1 {
         } else {
             Offer memory offer = offers[_collection][_offerId];
 
-            offers[_collection][offer.nextId].prevId = offer.prevId;
-            offers[_collection][offer.prevId].nextId = offer.nextId;
+            offers[_collection][offer.nextId].prevId = uint32(offer.prevId);
+            offers[_collection][offer.prevId].nextId = uint32(offer.nextId);
 
             delete offers[_collection][_offerId];
         }
@@ -262,7 +262,7 @@ contract CollectionOfferBookV1 {
     /// @notice Finds a collection offer to fill
     /// @param _collection The address of the ERC-721 collection
     /// @param _minAmount The minimum offer amount valid to match
-    function _getMatchingOffer(address _collection, uint256 _minAmount) internal view returns (uint32) {
+    function _getMatchingOffer(address _collection, uint256 _minAmount) internal view returns (uint256) {
         // If current ceiling offer is greater than or equal to maker's minimum, return its id to fill
         if (ceilingOfferAmount[_collection] >= _minAmount) {
             return ceilingOfferId[_collection];
@@ -283,21 +283,21 @@ contract CollectionOfferBookV1 {
     /// @notice Checks whether a given offer is the only one for a collection
     /// @param _collection The address of the ERC-721 collection
     /// @param _offerId The ID of the offer
-    function _isOnlyOffer(address _collection, uint32 _offerId) private view returns (bool) {
+    function _isOnlyOffer(address _collection, uint256 _offerId) private view returns (bool) {
         return (_offerId == floorOfferId[_collection]) && (_offerId == ceilingOfferId[_collection]);
     }
 
     /// @notice Checks whether a given offer is the collection ceiling
     /// @param _collection The address of the ERC-721 collection
     /// @param _offerId The ID of the offer
-    function _isCeilingOffer(address _collection, uint32 _offerId) private view returns (bool) {
+    function _isCeilingOffer(address _collection, uint256 _offerId) private view returns (bool) {
         return (_offerId == ceilingOfferId[_collection]);
     }
 
     /// @notice Checks whether a given offer is the collection floor
     /// @param _collection The address of the ERC-721 collection
     /// @param _offerId The ID of the offer
-    function _isFloorOffer(address _collection, uint32 _offerId) private view returns (bool) {
+    function _isFloorOffer(address _collection, uint256 _offerId) private view returns (bool) {
         return (_offerId == floorOfferId[_collection]);
     }
 
@@ -322,12 +322,12 @@ contract CollectionOfferBookV1 {
     /// @param _increase Whether the update is an amount increase or decrease
     function _isUpdateInPlace(
         address _collection,
-        uint32 _offerId,
+        uint256 _offerId,
         uint256 _newAmount,
         bool _increase
     ) private view returns (bool) {
-        uint32 nextOffer = offers[_collection][_offerId].nextId;
-        uint32 prevOffer = offers[_collection][_offerId].prevId;
+        uint256 nextOffer = offers[_collection][_offerId].nextId;
+        uint256 prevOffer = offers[_collection][_offerId].prevId;
         return
             ((_increase == true) && (_newAmount <= offers[_collection][nextOffer].amount)) ||
             ((_increase == false) && (_newAmount > offers[_collection][prevOffer].amount));
@@ -340,9 +340,9 @@ contract CollectionOfferBookV1 {
     /// @param _nextId The ID of the offer's next pointer
     function _connectNeighbors(
         address _collection,
-        uint32 _offerId,
-        uint32 _prevId,
-        uint32 _nextId
+        uint256 _offerId,
+        uint256 _prevId,
+        uint256 _nextId
     ) private {
         // If offer is floor --
         if (_offerId == floorOfferId[_collection]) {
@@ -363,8 +363,8 @@ contract CollectionOfferBookV1 {
             // Else offer is in middle --
         } else {
             // Update neighbor pointers
-            offers[_collection][_nextId].prevId = _prevId;
-            offers[_collection][_prevId].nextId = _nextId;
+            offers[_collection][_nextId].prevId = uint32(_prevId);
+            offers[_collection][_prevId].nextId = uint32(_nextId);
         }
     }
 
@@ -375,8 +375,8 @@ contract CollectionOfferBookV1 {
     /// @param _newAmount The new offer amount
     function _insertIncreasedOffer(
         address _collection,
-        uint32 _offerId,
-        uint32 _nextId,
+        uint256 _offerId,
+        uint256 _nextId,
         uint256 _newAmount
     ) private {
         Offer memory offer = offers[_collection][_nextId];
@@ -387,12 +387,12 @@ contract CollectionOfferBookV1 {
         }
 
         // Update offer pointers
-        offers[_collection][_offerId].nextId = offer.id;
-        offers[_collection][_offerId].prevId = offer.prevId;
+        offers[_collection][_offerId].nextId = uint32(offer.id);
+        offers[_collection][_offerId].prevId = uint32(offer.prevId);
 
         // Update neighbor pointers
-        offers[_collection][offer.id].prevId = _offerId;
-        offers[_collection][offer.prevId].nextId = _offerId;
+        offers[_collection][offer.id].prevId = uint32(_offerId);
+        offers[_collection][offer.prevId].nextId = uint32(_offerId);
 
         // Update offer amount
         offers[_collection][_offerId].amount = _newAmount;
@@ -405,8 +405,8 @@ contract CollectionOfferBookV1 {
     /// @param _newAmount The new offer amount
     function _insertDecreasedOffer(
         address _collection,
-        uint32 _offerId,
-        uint32 _prevId,
+        uint256 _offerId,
+        uint256 _prevId,
         uint256 _newAmount
     ) private {
         Offer memory offer = offers[_collection][_prevId];
@@ -417,12 +417,12 @@ contract CollectionOfferBookV1 {
         }
 
         // Update offer pointers
-        offers[_collection][_offerId].prevId = offer.id;
-        offers[_collection][_offerId].nextId = offer.nextId;
+        offers[_collection][_offerId].prevId = uint32(offer.id);
+        offers[_collection][_offerId].nextId = uint32(offer.nextId);
 
         // Update neighbor pointers
-        offers[_collection][offer.id].nextId = _offerId;
-        offers[_collection][offer.nextId].prevId = _offerId;
+        offers[_collection][offer.id].nextId = uint32(_offerId);
+        offers[_collection][offer.nextId].prevId = uint32(_offerId);
 
         // Update offer amount
         offers[_collection][_offerId].amount = _newAmount;
